Create Light_meter_retrieve_intensity.R
Light_meter_retrieve_intensity is a short script that pulls all the measured values from the .csv uMOL file that gets created when using the LICOR LI-180 spectrophotometer. For all the files in the folder, it creates a single OUTPUL.xlsx file with the index, filename and total light intensity. Subsequently, this can be used to make an overview of your LED light intensity.